Program.md
This program is made to be run in the Python language. The inputs for this program can be two things namely: Condition and Organization.
The conditions are the names of the diseases to be searched for (Cancer, Cholera, etc.) and the organizations are the companies that have done the studies (Pfizer, Johnson & Johnson, etc.).